mark more test to skip for python 3.7

+++ tests/test_running_services.py
@@ -40,12 +40,14 @@ class RealServiceTest(unittest.TestCase):
         """
         localdir = os.path.dirname(__file__)
         config = os.path.join(localdir, 'config.yaml')
-        cls.runner = subprocess.Popen(('nameko', 'run', '--config', config, 'services'), cwd=localdir)
-        time.sleep(1)
+        if not sys.version_info > (3, 6):
+            cls.runner = subprocess.Popen(('nameko', 'run', '--config', config, 'services'), cwd=localdir)
+            time.sleep(1)
 
     @classmethod
     def tearDownClass(cls):
-        cls.runner.kill()
+        if not sys.version_info > (3, 6):
+            cls.runner.kill()
 
     def test_echo_no_rpc(self):
         assert EchoService().echo(42) == (42,)
@@ -74,6 +76,7 @@ class RealServiceTest(unittest.TestCase):
 
         destroy_pool()
 
+    @unittest.skipIf(sys.version_info > (3, 6), "currently eventlet is broken on python 3.7+")
     @override_settings(NAMEKO_CONFIG=config)
     def test_pool_call_method_notdefined(self):
         with tools.assert_raises(MethodNotFound):
@@ -83,6 +86,7 @@ class RealServiceTest(unittest.TestCase):
 
         destroy_pool()
 
+    @unittest.skipIf(sys.version_info > (3, 6), "currently eventlet is broken on python 3.7+")
     @override_settings(NAMEKO_CONFIG={
         'AMQP_URI': 'amqp://guest:guest@localhost:6666'
     })
@@ -109,6 +113,7 @@ class RealServiceTest(unittest.TestCase):
 
         destroy_pool()
 
+    @unittest.skipIf(sys.version_info > (3, 6), "currently eventlet is broken on python 3.7+")
     @override_settings(NAMEKO_CONFIG=config)
     def test_pool_destroy_and_recreate(self):
         pool = get_pool()
@@ -122,6 +127,7 @@ class RealServiceTest(unittest.TestCase):
 
         destroy_pool()
 
+    @unittest.skipIf(sys.version_info > (3, 6), "currently eventlet is broken on python 3.7+")
     @override_settings(NAMEKO_CONFIG=config, NAMEKO_CONTEXT_DATA={"data": 123})
     def test_error_clear_context(self):
         pool = get_pool()
